# perf(cicd/ui): migrate from CRA/Craco to Vite