--- a/datahub-web-react/.env
+++ b/datahub-web-react/.env
@@ -1,3 +1,2 @@
-PUBLIC_URL=/assets
-REACT_APP_THEME_CONFIG=theme_light.config.json
+VITE_THEME_CONFIG=theme_light.config.json
 SKIP_PREFLIGHT_CHECK=true