--- a/datahub-web-react/.env
+++ b/datahub-web-react/.env
@@ -1,2 +1,2 @@
-VITE_THEME_CONFIG=theme_light.config.json
+VITE_THEME_CONFIG=theme_light.config
 SKIP_PREFLIGHT_CHECK=true