--- a/datahub-web-react/build.gradle
+++ b/datahub-web-react/build.gradle
@@ -84,13 +84,13 @@ configurations {
 distZip {
   dependsOn yarnBuild
   baseName 'datahub-web-react'
-  from 'dist'
+  from 'build'
 }
 
 task copyAssets {
   dependsOn distZip
   copy {
-    from 'dist'
+    from 'build'
     into '../datahub-frontend/public/'
   }
 }