--- a/datahub-web-react/package.json
+++ b/datahub-web-react/package.json
@@ -139,7 +139,6 @@
         "@typescript-eslint/eslint-plugin": "^4.25.0",
         "@typescript-eslint/parser": "^4.25.0",
         "@vitejs/plugin-react": "^2.2.0",
-        "babel-loader": "8.2.2",
         "eslint": "^7.27.0",
         "eslint-config-airbnb-typescript": "^12.3.1",
         "eslint-config-prettier": "^8.3.0",