--- a/datahub-web-react/package.json
+++ b/datahub-web-react/package.json
@@ -71,7 +71,6 @@
         "react": "^17.0.0",
         "react-color": "^2.19.3",
         "react-dom": "^17.0.0",
-        "react-email-share-link": "^1.0.3",
         "react-helmet": "^6.1.0",
         "react-highlighter": "^0.4.3",
         "react-icons": "4.3.1",