--- a/datahub-web-react/src/App.less
+++ b/datahub-web-react/src/App.less
@@ -1,4 +1,4 @@
-@import '~antd/dist/antd.less';
+@import 'antd/dist/antd.less';
 @import './conf/theme/global-overrides.less';
 
 @font-face {