--- a/datahub-web-react/src/App.tsx
+++ b/datahub-web-react/src/App.tsx
@@ -77,7 +77,7 @@ const App: React.VFC = () => {
     const [dynamicThemeConfig, setDynamicThemeConfig] = useState<Theme>(defaultThemeConfig);
 
     useEffect(() => {
-        import(`./conf/theme/${process.env.REACT_APP_THEME_CONFIG}`).then((theme) => {
+        import(/* @vite-ignore */ `./conf/theme/${import.meta.env.VITE_THEME_CONFIG}.json`).then((theme) => {
             setDynamicThemeConfig(theme);
         });
     }, []);