--- a/datahub-web-react/src/app/Routes.tsx
+++ b/datahub-web-react/src/app/Routes.tsx
@@ -37,8 +37,6 @@ export const Routes = (): JSX.Element => {
             <Route path={PageRoutes.SIGN_UP} component={SignUp} />
             <Route path={PageRoutes.RESET_CREDENTIALS} component={ResetCredentials} />
             <ProtectedRoute isLoggedIn={isLoggedIn} render={() => <ProtectedRoutes />} />
-            {/* Starting the react app locally opens /assets by default. For a smoother dev experience, we'll redirect to the homepage */}
-            <Route path={PageRoutes.ASSETS} component={() => <Redirect to="/" />} exact />
             <Route path="/*" component={NoPageFound} />
         </Switch>
     );