--- a/datahub-web-react/src/app/shared/share/items/EmailMenuItem.tsx
+++ b/datahub-web-react/src/app/shared/share/items/EmailMenuItem.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { Tooltip } from 'antd';
 import { CheckOutlined, MailOutlined } from '@ant-design/icons';
-import EmailShare from 'react-email-share-link';
 import MenuItem from 'antd/lib/menu/MenuItem';
 import { ANTD_GRAY } from '../../../entity/shared/constants';
 
@@ -24,7 +23,9 @@ export default function EmailMenuItem({ urn, name, type, key }: EmailMenuItemPro
      * Whether button has been clicked
      */
     const [isClicked, setIsClicked] = useState(false);
-    const linkText = window.location.href;
+    const subject = `${name} | ${type}`;
+    const body = `Check out this ${type} on DataHub: ${window.location.href}`;
+    const link = `mailto:?subject=${subject}&body=${body}`;
 
     return (
         <StyledMenuItem
@@ -34,18 +35,14 @@ export default function EmailMenuItem({ urn, name, type, key }: EmailMenuItemPro
                 setIsClicked(true);
             }}
         >
-            <EmailShare subject={`${name} | ${type}`} body={`Check out this ${type} on DataHub: ${linkText}`}>
-                {(link) => (
-                    <Tooltip title={`Share this ${type} via email`}>
-                        {isClicked ? <CheckOutlined /> : <MailOutlined />}
-                        <span>
-                            <a href={link} style={{ color: 'inherit' }}>
-                                <b>Email</b>
-                            </a>
-                        </span>
-                    </Tooltip>
-                )}
-            </EmailShare>
+            <Tooltip title={`Share this ${type} via email`}>
+                {isClicked ? <CheckOutlined /> : <MailOutlined />}
+                <span>
+                    <a href={link} style={{ color: 'inherit' }}>
+                        <b>Email</b>
+                    </a>
+                </span>
+            </Tooltip>
         </StyledMenuItem>
     );
 }