--- a/datahub-web-react/src/conf/theme/global-overrides.less
+++ b/datahub-web-react/src/conf/theme/global-overrides.less
@@ -1,3 +1,5 @@
+@import 'global-variables.less';
+
 element {
     --primary-color: brown;
 }