--- a/datahub-web-react/src/conf/theme/theme_dark.config.json
+++ b/datahub-web-react/src/conf/theme/theme_dark.config.json
@@ -18,7 +18,7 @@
         "homepage-background-lower-fade": "#333E4C"
     },
     "assets": {
-        "logoUrl": "/assets/logo.png"
+        "logoUrl": "/logo.png"
     },
     "content": {
         "title": "DataHub",