--- a/datahub-web-react/src/conf/theme/theme_light.config.json
+++ b/datahub-web-react/src/conf/theme/theme_light.config.json
@@ -21,7 +21,7 @@
         "box-shadow-hover": "0px 1px 0px 0.5px rgb(239 239 239)"
     },
     "assets": {
-        "logoUrl": "/assets/logo.png"
+        "logoUrl": "/logo.png"
     },
     "content": {
         "title": "DataHub",