--- a/datahub-web-react/src/graphql-mock/createServer.ts
+++ b/datahub-web-react/src/graphql-mock/createServer.ts
@@ -1,12 +1,10 @@
 /* eslint-disable global-require */
 /* eslint-disable @typescript-eslint/no-var-requires */
 
-if (process.env.REACT_APP_MOCK === 'true' || process.env.REACT_APP_MOCK === 'cy') {
-    if (process.env.REACT_APP_MOCK === 'cy') {
-        require('./server').makeServerForCypress();
-    } else {
-        require('./server').makeServer();
-    }
+if (import.meta.env.VITE_MOCK === 'true') {
+    require('./server').makeServer();
+} else if (import.meta.env.VITE_MOCK === 'cy') {
+    require('./server').makeServerForCypress();
 }
 
 export {};