--- a/datahub-web-react/src/vite-env.d.ts
+++ b/datahub-web-react/src/vite-env.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="vite/client" />
+/// <reference types="vite-plugin-svgr/client" />