--- a/datahub-web-react/vite.config.ts
+++ b/datahub-web-react/vite.config.ts
@@ -29,6 +29,9 @@ export default defineConfig({
         open: true,
         port: 3000,
     },
+    optimizeDeps: {
+        exclude: ['babel-runtime'],
+    },
     css: {
         preprocessorOptions: {
             less: {