--- a/metadata-ingestion/examples/mce_files/data_platforms.json
+++ b/metadata-ingestion/examples/mce_files/data_platforms.json
@@ -11,7 +11,7 @@
               "name": "adlsGen1",
               "displayName": "Azure Data Lake (Gen 1)",
               "type": "FILE_SYSTEM",
-              "logoUrl": "/assets/platforms/adlslogo.png"
+              "logoUrl": "/platforms/adlslogo.png"
             }
           }
         ]
@@ -31,7 +31,7 @@
               "name": "adlsGen2",
               "displayName": "Azure Data Lake (Gen 2)",
               "type": "FILE_SYSTEM",
-              "logoUrl": "/assets/platforms/adlslogo.png"
+              "logoUrl": "/platforms/adlslogo.png"
             }
           }
         ]
@@ -70,7 +70,7 @@
               "name": "couchbase",
               "displayName": "Couchbase",
               "type": "KEY_VALUE_STORE",
-              "logoUrl": "/assets/platforms/couchbaselogo.png"
+              "logoUrl": "/platforms/couchbaselogo.png"
             }
           }
         ]
@@ -109,7 +109,7 @@
               "name": "hdfs",
               "displayName": "HDFS",
               "type": "FILE_SYSTEM",
-              "logoUrl": "/assets/platforms/hadooplogo.png"
+              "logoUrl": "/platforms/hadooplogo.png"
             }
           }
         ]
@@ -129,7 +129,7 @@
               "name": "hive",
               "displayName": "Hive",
               "type": "FILE_SYSTEM",
-              "logoUrl": "/assets/platforms/hivelogo.png"
+              "logoUrl": "/platforms/hivelogo.png"
             }
           }
         ]
@@ -149,7 +149,7 @@
               "name": "s3",
               "displayName": "AWS S3",
               "type": "FILE_SYSTEM",
-              "logoUrl": "/assets/platforms/s3.png"
+              "logoUrl": "/platforms/s3.png"
             }
           }
         ]
@@ -169,7 +169,7 @@
               "name": "kafka",
               "displayName": "Kafka",
               "type": "MESSAGE_BROKER",
-              "logoUrl": "/assets/platforms/kafkalogo.png"
+              "logoUrl": "/platforms/kafkalogo.png"
             }
           }
         ]
@@ -189,7 +189,7 @@
               "name": "kusto",
               "displayName": "Kusto",
               "type": "OLAP_DATASTORE",
-              "logoUrl": "/assets/platforms/kustologo.png"
+              "logoUrl": "/platforms/kustologo.png"
             }
           }
         ]
@@ -209,7 +209,7 @@
               "name": "mode",
               "displayName": "Mode",
               "type": "KEY_VALUE_STORE",
-              "logoUrl": "/assets/platforms/modelogo.png"
+              "logoUrl": "/platforms/modelogo.png"
             }
           }
         ]
@@ -229,7 +229,7 @@
               "name": "mongodb",
               "displayName": "MongoDB",
               "type": "KEY_VALUE_STORE",
-              "logoUrl": "/assets/platforms/mongodblogo.png"
+              "logoUrl": "/platforms/mongodblogo.png"
             }
           }
         ]
@@ -249,7 +249,7 @@
               "name": "mysql",
               "displayName": "MySQL",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/mysqllogo.png"
+              "logoUrl": "/platforms/mysqllogo.png"
             }
           }
         ]
@@ -269,7 +269,7 @@
               "name": "oracle",
               "displayName": "Oracle",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/oraclelogo.png"
+              "logoUrl": "/platforms/oraclelogo.png"
             }
           }
         ]
@@ -289,7 +289,7 @@
               "name": "pinot",
               "displayName": "Pinot",
               "type": "OLAP_DATASTORE",
-              "logoUrl": "/assets/platforms/pinotlogo.png"
+              "logoUrl": "/platforms/pinotlogo.png"
             }
           }
         ]
@@ -309,7 +309,7 @@
               "name": "postgres",
               "displayName": "PostgreSQL",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/postgreslogo.png"
+              "logoUrl": "/platforms/postgreslogo.png"
             }
           }
         ]
@@ -329,7 +329,7 @@
               "name": "presto",
               "displayName": "Presto",
               "type": "QUERY_ENGINE",
-              "logoUrl": "/assets/platforms/prestologo.png"
+              "logoUrl": "/platforms/prestologo.png"
             }
           }
         ]
@@ -349,7 +349,7 @@
               "name": "teradata",
               "displayName": "Teradata",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/teradatalogo.png"
+              "logoUrl": "/platforms/teradatalogo.png"
             }
           }
         ]
@@ -388,7 +388,7 @@
               "name": "snowflake",
               "displayName": "Snowflake",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/snowflakelogo.png"
+              "logoUrl": "/platforms/snowflakelogo.png"
             }
           }
         ]
@@ -408,7 +408,7 @@
               "name": "superset",
               "displayName": "Superset",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/supersetlogo.png"
+              "logoUrl": "/platforms/supersetlogo.png"
             }
           }
         ]
@@ -428,7 +428,7 @@
               "name": "datahub",
               "displayName": "DataHub",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/datahublogo.png"
+              "logoUrl": "/platforms/datahublogo.png"
             }
           }
         ]
@@ -448,7 +448,7 @@
               "name": "redshift",
               "displayName": "Redshift",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/redshiftlogo.png"
+              "logoUrl": "/platforms/redshiftlogo.png"
             }
           }
         ]
@@ -468,7 +468,7 @@
               "name": "mssql",
               "displayName": "SQL Server",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/mssqllogo.png"
+              "logoUrl": "/platforms/mssqllogo.png"
             }
           }
         ]
@@ -488,7 +488,7 @@
               "name": "bigquery",
               "displayName": "BigQuery",
               "type": "RELATIONAL_DB",
-              "logoUrl": "/assets/platforms/bigquerylogo.png"
+              "logoUrl": "/platforms/bigquerylogo.png"
             }
           }
         ]
@@ -508,7 +508,7 @@
               "name": "druid",
               "displayName": "Druid",
               "type": "OLAP_DATASTORE",
-              "logoUrl": "/assets/platforms/druidlogo.png"
+              "logoUrl": "/platforms/druidlogo.png"
             }
           }
         ]
@@ -528,7 +528,7 @@
               "name": "looker",
               "displayName": "Looker",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/lookerlogo.png"
+              "logoUrl": "/platforms/lookerlogo.png"
             }
           }
         ]
@@ -548,7 +548,7 @@
               "name": "feast",
               "displayName": "Feast",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/feastlogo.png"
+              "logoUrl": "/platforms/feastlogo.png"
             }
           }
         ]
@@ -568,7 +568,7 @@
               "name": "sagemaker",
               "displayName": "SageMaker",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/sagemakerlogo.png"
+              "logoUrl": "/platforms/sagemakerlogo.png"
             }
           }
         ]
@@ -588,7 +588,7 @@
               "name": "glue",
               "displayName": "Glue",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/gluelogo.png"
+              "logoUrl": "/platforms/gluelogo.png"
             }
           }
         ]
@@ -608,7 +608,7 @@
               "name": "redash",
               "displayName": "Redash",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/redashlogo.png"
+              "logoUrl": "/platforms/redashlogo.png"
             }
           }
         ]
@@ -628,7 +628,7 @@
               "name": "tableau",
               "displayName": "Tableau",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/tableaulogo.png"
+              "logoUrl": "/platforms/tableaulogo.png"
             }
           }
         ]
@@ -648,7 +648,7 @@
               "name": "powerbi",
               "displayName": "Power BI",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/powerbilogo.png"
+              "logoUrl": "/platforms/powerbilogo.png"
             }
           }
         ]
@@ -668,7 +668,7 @@
               "name": "powerbi-report-server",
               "displayName": "Power BI Report Server",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/powerbireportserverlogo.png"
+              "logoUrl": "/platforms/powerbireportserverlogo.png"
             }
           }
         ]
@@ -687,7 +687,7 @@
               "name": "Great Expectations",
               "displayName": "Great Expectations",
               "type": "OTHERS",
-              "logoUrl": "/assets/platforms/greatexpectationslogo.png",
+              "logoUrl": "/platforms/greatexpectationslogo.png",
               "datasetNameDelimiter": "."
             }
           }