--- a/metadata-service/factories/src/main/resources/application.yml
+++ b/metadata-service/factories/src/main/resources/application.yml
@@ -92,8 +92,8 @@ platformAnalytics:
 
 visualConfig:
   assets:
-    logoUrl: ${REACT_APP_LOGO_URL:/assets/platforms/datahublogo.png}
-    faviconUrl: ${REACT_APP_FAVICON_URL:/assets/favicon.ico}
+    logoUrl: ${VITE_LOGO_URL:/platforms/datahublogo.png}
+    faviconUrl: ${VITE_FAVICON_URL:/favicon.ico}
 
 # Storage Layer