## Overview

This migration from the current create-react-app build tooling is a pretty significant overhaul that touches quite a few areas. The primary reason for wanting to migrate is quicker dev server startup and production builds. It's also worth noting that the creator/lead maintainer of Webpack has stepped away from the project to work on a new build tool ([Turbopack](https://vercel.com/blog/turbopack)), causing some uncertainty for the long-term project roadmap as others step in to maintain.

## Benchmarks

**Dev server startup** (`yarn start`)
- Baseline:   52.74s
- With Vite: 0.83s

**Production build** (`yarn build`)
- Baseline: 96.95s
- With Vite: 37.52s

All tests were performed using the latest from master (as of 11/30/22) on a 2021 M1 MBP (32GB). Average time was taken from 3 consecutive runs of the same test. Time spent generating graphql types (`yarn generate`) is excluded from comparison.

## Major Changes
- Local dev proxy server rewritten as Vite plugin (see `devProxyPlugin.ts`)
- Frontend build pipeline no longer moves static assets from `build/` to `dist/`. Vite outputs to dist by default.
- Removal of manually prepending `assets/` to certain urls. Vite should handle this automatically when hosting files between local dev and production builds.

## What's left?
- [ ] Fix Cypress startup
- [ ] Fix remaining asset 404s (monaco editor library 3rd party JS and platform logos)

## Checklist

- [ ] The PR conforms to DataHub's [Contributing Guideline](https://github.com/datahub-project/datahub/blob/master/docs/CONTRIBUTING.md) (particularly [Commit Message Format](https://github.com/datahub-project/datahub/blob/master/docs/CONTRIBUTING.md#commit-message-format))
- [ ] Links to related issues (if applicable)
- [ ] Tests for the changes have been added/updated (if applicable)
- [ ] Docs related to the changes have been added/updated (if applicable). If a new feature has been added a Usage Guide has been added for the same.
- [ ] For any breaking change/potential downtime/deprecation/big changes an entry has been made in [Updating DataHub](https://github.com/datahub-project/datahub/blob/master/docs/how/updating-datahub.md)