@stanbaker 2023-01-04
> hey @stanbaker! Thanks so much for putting this together. I'm curious how important a change like this is for your team? In your description you mention the build times being the primary reason for the change, as well as some uncertainty about CRA's future leadership. How much of a blocker are build times for your team or is this just a nice to have?
> 
> The core team has some concerns about the risks of switching away from CRA due to all of the hidden complexities with single-page app builds and configurations that it takes care of for us. Also, while Vite certainly has a large following and usage, it's still somewhat less established than create-react-app (I feel less strongly about this than the risks of making the switch). Basically, we want to be very cautious and make sure we're doing the right thing here!
> 
> I think we're leaning towards holding off on this move until we have more signal that it's the right call, but we'd love to hear some more thoughts from you and your team.

Hey @chriscollins3456, thanks for the feedback! This isn't a blocker/priority by any means, totally understand if you all would rather stick with what's working for now instead of taking some risk on entirely migrating build tooling! Closing this PR